vdev: fix sign compare warnings.
E.g.
cc -D_GNU_SOURCE -DLOG_VERBOSITY=LOG_INFO -I./include -I. -Wall -Werror -Wunused-parameter -Wsign-compare  -o vdev.o -c vdev.c
vdev.c: In function ‘vhd_gpa_range_mark_dirty’:
vdev.c:412:47: error: comparison of integer expressions of different signedness: ‘uint64_t’ {aka ‘long unsigned int’} and ‘int’ [-Werror=sign-compare]
  412 |         for (; page <= last_page && page / 64 == chunk; ++page) {
      |

+++ vdev.c
@@ -152,7 +152,7 @@ static int net_send_msg_fds(int fd, const struct vhost_user_msg *msg,
     if (len < 0) {
         VHD_LOG_ERROR("sendmsg() failed: %d", errno);
         return -errno;
-    } else if (len != (VHOST_MSG_HDR_SIZE + msg->size)) {
+    } else if ((unsigned)len != (VHOST_MSG_HDR_SIZE + msg->size)) {
         VHD_LOG_ERROR("sendmsg() puts less bytes = %d, than required = %lu",
                 len, VHOST_MSG_HDR_SIZE + msg->size);
         return -EIO;
@@ -408,7 +408,7 @@ void vhd_gpa_range_mark_dirty(struct vhd_guest_memory_map *mm,
      */
     do {
         uint64_t mask = 0UL;
-        int chunk = page / 64;
+        uint64_t chunk = page / 64;
         for (; page <= last_page && page / 64 == chunk; ++page) {
             mask |= 1LU << (page % 64);
         }
